// src/generated/live/PlaySession.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SessionSegment } from "./SessionSegment";

export type PlaySession = { id: string, title: string, startedAt: number, endedAt: number | null, endReason: string | null, segments: Array<SessionSegment>, };